fix: restore the loading skeleton on navigation to the Decision Instance route
The route's loader awaits the decision instance fetch before mounting the page, so
the page-component-level skeleton (status 'pending') was unreachable on a real
navigation - the query is already resolved by the time the component renders. Add a
route-level pendingComponent that renders the same InstanceHeaderSkeleton, sharing
getHeaderColumns (extracted to its own module to satisfy react-refresh/only-export-
components) and a new DecisionInstanceShell so the loading and loaded views stay
in sync.
